+++ OpticEMS.Contracts/Services/SignalPreprocessing/TrendResult.cs
@@ -2,8 +2,8 @@
 {
     public class TrendResult()
     {
-        public uint[] FrameAveraged { get; init; } = Array.Empty<uint>();
-        public uint[] Smoothed { get; init; } = Array.Empty<uint>();
+        public double[] FrameAveraged { get; init; } = Array.Empty<double>();
+        public double[] Smoothed { get; init; } = Array.Empty<double>();
         public double[] Derivatives { get; init; } = Array.Empty<double>();
         public double Timestamp { get; init; }
