+++ OpticEMS.Notifications/Messages/CalibrationChartUpdatedMessage.cs
@@ -1,4 +1,4 @@
 ﻿namespace OpticEMS.Notifications.Messages
 {
-    public record CalibrationChartUpdatedMessage(int Id, uint[] Intensities);
+    public record CalibrationChartUpdatedMessage(int Id, double[] Intensities);
 }
