+++ OpticEMS.Notifications/Messages/LiveSpectrumDataMessage.cs
@@ -1,4 +1,4 @@
 ﻿namespace OpticEMS.Notifications.Messages
 {
-    public record LiveSpectrumDataMessage(int ChannelId, double[] Wavelengths, uint[] Intensities);
+    public record LiveSpectrumDataMessage(int ChannelId, double[] Wavelengths, double[] Intensities);
 }
