+++ OpticEMS.Notifications/Messages/SpectrumUpdatedMessage.cs
@@ -1,4 +1,4 @@
 ﻿namespace OpticEMS.Notifications.Messages
 {
-    public record SpectrumUpdatedMessage(int ChannelId, uint[] Intensities, double[] Wavelengths);
+    public record SpectrumUpdatedMessage(int ChannelId, double[] Intensities, double[] Wavelengths);
 }
